requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.2
pandas==2.2.3
pillow==11.0.0
PyAudio==0.2.14
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
SpeechRecognition==3.11.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
